animalcules: Interactive microbiome analysis toolkit

animalcules is an R package for utilizing up-to-date data analytics, visualization methods, and machine learning models to provide users an easy-to-use interactive microbiome analysis framework. It can be used as a standalone software package or users can explore their data with the accompanying interactive R Shiny application. Traditional microbiome analysis such as alpha/beta diversity and differential abundance analysis are enhanced, while new methods like biomarker identification are introduced by animalcules. Powerful interactive and dynamic figures generated by animalcules enable users to understand their data better and discover new insights.

See also

Author

Maintainer: Yue Zhao yuezh@bu.edu (ORCID)

Authors:

Anthony Federico anfed@bu.edu (ORCID)

W. Evan Johnson wej@bu.edu (ORCID)