Reads the data from PathoScope reports and returns a list of final guess relative abundance and count data

Reads the data from PathoScope reports and returns a list of final guess relative abundance and count data

read_pathoscope_data(
  input_dir = ".",
  pathoreport_file_suffix = "-sam-report.tsv",
  use.input.files = FALSE,
  input.files.path.vec = NULL,
  input.files.name.vec = NULL
)

Arguments

input_dir

Directory where the tsv files from PathoScope are located

pathoreport_file_suffix

PathoScope report files suffix

use.input.files

whether input dir to pathoscope files or directly pathoscope files

input.files.path.vec

vector of pathoscope file paths

input.files.name.vec

vector of pathoscope file names

Value

List of final guess relative abundance and count data