Dimensionality reduction through t-SNE

Dimensionality reduction through t-SNE

dimred_tsne(MAE, tax_level, color, shape = NULL, k = c("2D", "3D"),
  initial_dims = 30, perplexity = 10, datatype = c("logcpm",
  "relabu", "counts"), tsne_cache = NULL)

Arguments

MAE
A multi-assay experiment object

tax_level
The taxon level used for organisms

color
A condition to color data points by e.g. "AGE"

shape
A condition to shape data points by e.g. "SEX"

k
Plot dimensions e.g. c("2D","3D")

initial_dims
The number of dimensions to use in reduction method

perplexity
Optimal number of neighbors

datatype
Datatype to use e.g. c("logcpm", "relabu", "counts")

tsne_cache
Pass the cached data back into the function

Value

A list with a plotly object and cached data

Examples

data_dir = system.file("extdata/MAE.rds", package = "animalcules")
toy_data <- readRDS(data_dir)
p <- dimred_tsne(toy_data,
                 tax_level="phylum",
                 color="AGE",
                 shape="GROUP",
                 k="3D",
                 initial_dims=30,
                 perplexity=10,
                 datatype="logcpm")

#> sigma summary: Min. : 0.439403562593612 |1st Qu. : 0.614507809712872 |Median : 0.678538111578798 |Mean : 0.677688154944898 |3rd Qu. : 0.742449349451526 |Max. : 0.916738699643141 |

#> Epoch: Iteration #100 error is: 17.3254624174099

#> Epoch: Iteration #200 error is: 1.10221891600239

#> Epoch: Iteration #300 error is: 0.669765639361604

#> Epoch: Iteration #400 error is: 0.481946218890382

#> Epoch: Iteration #500 error is: 0.314989814726309

#> Epoch: Iteration #600 error is: 0.28029745232966

#> Epoch: Iteration #700 error is: 0.272248021007063

#> Epoch: Iteration #800 error is: 0.264591762293109

#> Epoch: Iteration #900 error is: 0.26038780252

#> Epoch: Iteration #1000 error is: 0.259517352548058

p

#> $plot
#> 
#> $data
#>                     X           Y            Z
#> subject_1   -93.79049  -90.590755 -149.8324092
#> subject_2   -58.78947 -135.526235 -123.1239464
#> subject_3   -47.88005 -297.416106 -259.0735899
#> subject_4   285.54884   42.891286   99.4293559
#> subject_5   219.97810   40.431065   53.9742712
#> subject_6  -166.80371   41.589405  101.3328644
#> subject_7   -56.13490  -27.753980  204.8088764
#> subject_8   -26.25678  -40.077947   68.5394450
#> subject_9  -115.33578  -57.678079   45.8556971
#> subject_10  -59.78437 -138.244253  173.7078721
#> subject_11  -26.31887 -137.887646  228.6760136
#> subject_12  209.71313   79.893616   23.0178264
#> subject_13  -71.44406  100.027104  454.3822904
#> subject_14   80.81051   51.161266   97.9103081
#> subject_15  285.46516 -228.707421 -153.2634818
#> subject_16   26.83287  274.850465  100.4637204
#> subject_17 -163.86542  196.948831 -198.9661855
#> subject_18  244.56265  232.698422 -227.6007988
#> subject_19  262.50406 -264.361846 -174.7268368
#> subject_20  225.72992  192.590504 -213.5880384
#> subject_21 -184.05130   67.837373 -122.1181217
#> subject_22  -44.18571  138.420106  445.7159854
#> subject_23   11.36840 -154.818438   15.2363830
#> subject_24 -141.32190  236.380325   18.4395702
#> subject_25   27.20246 -207.145444   39.4301674
#> subject_26 -163.11214  133.801111 -102.4022140
#> subject_27  -72.52986 -264.243964 -229.9944366
#> subject_28 -167.83150  163.139126 -147.9951265
#> subject_29   10.25089 -196.897847 -111.0142201
#> subject_30  222.02002 -379.449678   92.5115214
#> subject_31  -30.23696  135.961758 -303.7868959
#> subject_32  -91.65629 -201.292314  157.2717579
#> subject_33 -120.50369  200.579740 -265.4766097
#> subject_34  -93.05029  128.144763  537.7214534
#> subject_35 -126.26651  125.970671  452.7827024
#> subject_36 -149.22175  282.618633    0.8866505
#> subject_37  -74.41525  256.836625   71.3575810
#> subject_38  -20.69591   14.895964   62.2966230
#> subject_39 -114.09287 -180.890022 -163.4255775
#> subject_40  178.73294    1.675347  -63.9076595
#> subject_41 -127.70366   -4.642105   66.6431994
#> subject_42 -149.85238  -10.205194 -158.2471367
#> subject_43   17.48320   90.752129 -273.1848383
#> subject_44   27.61062  226.138737   76.5603994
#> subject_45  202.63511 -388.329289   47.2030073
#> subject_46   57.13788   30.266747 -145.4384789
#> subject_47   88.82601   89.584620   63.6559686
#> subject_48   89.19547   64.290394 -180.2233843
#> subject_49 -122.06454 -266.751559   84.8473773
#> subject_50  105.58818   32.533990 -117.2689019
#>